+++ source/Analysis/Autocorrelation.vb
@@ -199,7 +199,7 @@ Friend Class Autocorrelation
 
         'Ergebnisdiagramm
         Me.ResultChart = New Steema.TeeChart.Chart()
-        Call Helpers.FormatChart(Me.ResultChart)
+        Call Helpers.ChartSetDefaultFormat(Me.ResultChart)
         Me.ResultChart.Header.Text = "Autocorrelation for " & ts_in.Title
 
         'X-Achse