+++ source/Dialogs/SettingsDialog.Designer.vb
@@ -32,10 +32,13 @@ Partial Class SettingsDialog
         Me.GroupBox_Chart = New System.Windows.Forms.GroupBox()
         Me.GroupBox_Logging = New System.Windows.Forms.GroupBox()
         Me.CheckBox_logShowDebugMessages = New System.Windows.Forms.CheckBox()
+        Me.Label_DefaultFontSize = New System.Windows.Forms.Label()
+        Me.NumericUpDown_DefaultFontSize = New System.Windows.Forms.NumericUpDown()
         Me.TableLayoutPanel1.SuspendLayout()
         CType(Me.NumericUpDown_DefaultLineWidth, System.ComponentModel.ISupportInitialize).BeginInit()
         Me.GroupBox_Chart.SuspendLayout()
         Me.GroupBox_Logging.SuspendLayout()
+        CType(Me.NumericUpDown_DefaultFontSize, System.ComponentModel.ISupportInitialize).BeginInit()
         Me.SuspendLayout()
         '
         'TableLayoutPanel1
@@ -77,7 +80,7 @@ Partial Class SettingsDialog
         '
         Me.NumericUpDown_DefaultLineWidth.Anchor = CType(((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
-        Me.NumericUpDown_DefaultLineWidth.Location = New System.Drawing.Point(133, 19)
+        Me.NumericUpDown_DefaultLineWidth.Location = New System.Drawing.Point(133, 45)
         Me.NumericUpDown_DefaultLineWidth.Maximum = New Decimal(New Integer() {10, 0, 0, 0})
         Me.NumericUpDown_DefaultLineWidth.Minimum = New Decimal(New Integer() {1, 0, 0, 0})
         Me.NumericUpDown_DefaultLineWidth.Name = "NumericUpDown_DefaultLineWidth"
@@ -88,7 +91,7 @@ Partial Class SettingsDialog
         'Label_defaultLineWidth
         '
         Me.Label_defaultLineWidth.AutoSize = True
-        Me.Label_defaultLineWidth.Location = New System.Drawing.Point(6, 21)
+        Me.Label_defaultLineWidth.Location = New System.Drawing.Point(6, 47)
         Me.Label_defaultLineWidth.Name = "Label_defaultLineWidth"
         Me.Label_defaultLineWidth.Size = New System.Drawing.Size(121, 13)
         Me.Label_defaultLineWidth.TabIndex = 2
@@ -107,11 +110,13 @@ Partial Class SettingsDialog
         '
         Me.GroupBox_Chart.Anchor = CType(((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
+        Me.GroupBox_Chart.Controls.Add(Me.NumericUpDown_DefaultFontSize)
+        Me.GroupBox_Chart.Controls.Add(Me.Label_DefaultFontSize)
         Me.GroupBox_Chart.Controls.Add(Me.Label_defaultLineWidth)
         Me.GroupBox_Chart.Controls.Add(Me.NumericUpDown_DefaultLineWidth)
         Me.GroupBox_Chart.Location = New System.Drawing.Point(12, 12)
         Me.GroupBox_Chart.Name = "GroupBox_Chart"
-        Me.GroupBox_Chart.Size = New System.Drawing.Size(210, 54)
+        Me.GroupBox_Chart.Size = New System.Drawing.Size(210, 78)
         Me.GroupBox_Chart.TabIndex = 5
         Me.GroupBox_Chart.TabStop = False
         Me.GroupBox_Chart.Text = "Chart"
@@ -122,7 +127,7 @@ Partial Class SettingsDialog
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
         Me.GroupBox_Logging.Controls.Add(Me.Label_loggingLevel)
         Me.GroupBox_Logging.Controls.Add(Me.CheckBox_logShowDebugMessages)
-        Me.GroupBox_Logging.Location = New System.Drawing.Point(12, 73)
+        Me.GroupBox_Logging.Location = New System.Drawing.Point(12, 96)
         Me.GroupBox_Logging.Name = "GroupBox_Logging"
         Me.GroupBox_Logging.Size = New System.Drawing.Size(210, 52)
         Me.GroupBox_Logging.TabIndex = 6
@@ -138,6 +143,27 @@ Partial Class SettingsDialog
         Me.CheckBox_logShowDebugMessages.Size = New System.Drawing.Size(15, 14)
         Me.CheckBox_logShowDebugMessages.TabIndex = 7
         Me.CheckBox_logShowDebugMessages.UseVisualStyleBackColor = True
+        '
+        'Label_DefaultFontSize
+        '
+        Me.Label_DefaultFontSize.AutoSize = True
+        Me.Label_DefaultFontSize.Location = New System.Drawing.Point(6, 21)
+        Me.Label_DefaultFontSize.Name = "Label_DefaultFontSize"
+        Me.Label_DefaultFontSize.Size = New System.Drawing.Size(86, 13)
+        Me.Label_DefaultFontSize.TabIndex = 3
+        Me.Label_DefaultFontSize.Text = "Default font size:"
+        '
+        'NumericUpDown_DefaultFontSize
+        '
+        Me.NumericUpDown_DefaultFontSize.Anchor = CType(((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Left) _
+            Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
+        Me.NumericUpDown_DefaultFontSize.Location = New System.Drawing.Point(133, 19)
+        Me.NumericUpDown_DefaultFontSize.Maximum = New Decimal(New Integer() {20, 0, 0, 0})
+        Me.NumericUpDown_DefaultFontSize.Minimum = New Decimal(New Integer() {6, 0, 0, 0})
+        Me.NumericUpDown_DefaultFontSize.Name = "NumericUpDown_DefaultFontSize"
+        Me.NumericUpDown_DefaultFontSize.Size = New System.Drawing.Size(71, 20)
+        Me.NumericUpDown_DefaultFontSize.TabIndex = 4
+        Me.NumericUpDown_DefaultFontSize.Value = New Decimal(New Integer() {10, 0, 0, 0})
         '
         'SettingsDialog
         '
@@ -164,6 +190,7 @@ Partial Class SettingsDialog
         Me.GroupBox_Chart.PerformLayout()
         Me.GroupBox_Logging.ResumeLayout(False)
         Me.GroupBox_Logging.PerformLayout()
+        CType(Me.NumericUpDown_DefaultFontSize, System.ComponentModel.ISupportInitialize).EndInit()
         Me.ResumeLayout(False)
 
     End Sub
@@ -176,4 +203,6 @@ Partial Class SettingsDialog
     Friend WithEvents GroupBox_Chart As GroupBox
     Friend WithEvents GroupBox_Logging As GroupBox
     Friend WithEvents CheckBox_logShowDebugMessages As CheckBox
+    Friend WithEvents NumericUpDown_DefaultFontSize As NumericUpDown
+    Friend WithEvents Label_DefaultFontSize As Label
 End Class