+++ source/My Project/Settings.Designer.vb
@@ -102,6 +102,18 @@ Namespace My
                 Me("MRUFiles") = value
             End Set
         End Property
+        
+        <Global.System.Configuration.UserScopedSettingAttribute(),  _
+         Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
+         Global.System.Configuration.DefaultSettingValueAttribute("10")>  _
+        Public Property defaultFontSize() As Integer
+            Get
+                Return CType(Me("defaultFontSize"),Integer)
+            End Get
+            Set
+                Me("defaultFontSize") = value
+            End Set
+        End Property
     End Class
 End Namespace